+++ packages/terra-core-docs/src/terra-dev-site/test/overlay/LoadingOverlay/CustomMessageLoadingOverlay.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoadingOverlay from '../../../../LoadingOverlay';
+import LoadingOverlay from 'terra-overlay/lib/LoadingOverlay';
 
 const CustomMessageLoadingOverlay = () => (
   <LoadingOverlay isOpen message="Custom Loading Message" id="terra-LoadingOverlay" />