+++ packages/terra-overlay/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Updated dependencies to not be hard-coded.
+  * Migrate docs to terra-core-docs.
 
 ## 3.69.0 - (November 2, 2021)